delete and crud operation

+++ src/app/membre-list/membre-list.component.ts
@@ -1,8 +1,11 @@
 
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { Member } from 'src/Models/member.model';
 import { MembersService } from 'src/Services/members.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 
 @Component({
@@ -13,25 +16,50 @@ import { MembersService } from 'src/Services/members.service';
 export class MembreListComponent implements OnInit {
 
   //any : quelque soit le type
-  public dataSource: Member[];
+  public dataSource: MatTableDataSource<Member>;
   //3al 9ad ma3andik columns tzidou fi displayedcolumns ==9adeh 3andik min ngcontainer fil html
-  displayedColumns: string[] = ['id', 'cin', 'name', 'createdDate', 'cv', 'type', 'Actions'];
+  displayedColumns: string[] = ['cin', 'nom', 'dateNaissance', 'cv', 'prenom', 'Actions'];
 
   //fil constructeur na3mel instance min il service : ma3neha injectit il service 
-  constructor(private ms: MembersService, private router: Router) {
-    this.dataSource = ms.tab
+  constructor(private ms: MembersService, private router: Router, private dialog: MatDialog) {
+    this.dataSource = new MatTableDataSource(this.ms.tab);
   }
   delete(id: string) {
-    this.ms.deleteMember(id).then(() => {
-      this.dataSource = this.ms.res;
-      console.log(this.dataSource)
-    });
 
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {})
+    // nlanci thread de type observable :no93ed netssana mba3d massakarha il user w nestana il retour de la type boolean 
+    dialogRef.afterClosed().subscribe(
+      isDeleted => {
+        if (isDeleted) {
+          //exécute de code de la suppression 
+          this.ms.deleteMember(id).then(() => {
+            this.GetMembers();
+
+          });
+
+        }
+      }
+
+    )
+  }
+  GetMembers(): void {
+    //.then((awelparamétre houwa chnou jek min il resolve)=>{chhnou na3mel a3lih})
+    //this.ms.GetAllMembers().then((data) => {
+    //.data  accéder au données de table de type MatTableDataSource
+    //this.dataSource.data = data
+    // })
+    this.ms.GetAllMembers();
+  }
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
 
 
   }
 
+
   ngOnInit(): void {
+    this.GetMembers();
   }
 
 }
